// Frontend/src/types/search.d.ts
// src/types.ts

export interface School {
    Adresa: string;
    Faks: string;
    Mjesto: string;
    Naziv: string;
    Osnivač: string;
    PoštanskiBroj: string;
    Ravnatelj: string;
    Telefon: string;
    TipUstanove: string;
    Web: string;
    _id: string;
    ePošta: string;
    Šifra: string;
    Županija: string;
  }
  